functions/requirements.txt
firebase-functions>=0.1.0
google-cloud-firestore>=2.5.0
google-cloud-storage>=2.1.0
google-cloud-tasks>=2.7.0
google-cloud-aiplatform>=1.24.0
Flask>=2.0.0
python-dateutil>=2.8.2
requests>=2.25.0
